wireguard: device: fill in skb->protocol for zero case
One way to handle AF_PACKET would be to implement
header_ops->parse_protocol, which AF_PACKET calls via:

    packet_sendmsg -> packet_snd -> packet_parse_headers
      -> dev_parse_header_protocol -> parse_protocol

This works, but it seems like various other behavior changes simply on
the basis of whether or not dev->header_ops == NULL or not. It's easier,
therefore, to just fill in skb->protocol when it's zero inside the xmit
function, and not change the packet flow otherwise. So this commit
checks for skb->protocol==0, and sets it in that case.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>